requirements.txt
pywin32>=305
Pillow>=10.0.0

